feat(projects): enable feature flag to redirect users to project link instead project page
